moved time series knn into distanced based submodule
