+++ sktime/classifiers/distance_based/__init__.py
@@ -2,3 +2,4 @@ from sktime.classifiers.distance_based.elastic_ensemble import ElasticEnsemble
 from sktime.classifiers.distance_based.proximity_forest import ProximityForest
 from sktime.classifiers.distance_based.proximity_forest import ProximityStump
 from sktime.classifiers.distance_based.proximity_forest import ProximityTree
+from sktime.classifiers.distance_based.time_series_neighbors import KNeighborsTimeSeriesClassifier
